Exclude `/[repo name]` from link checker
When published on github pages, the root of the
site will be at `/[repo name]` e.g. 
`/test-doc-site`

There is no corresponding HTML file in the 
`./docs` directory (it's the `./docs/index.html`
file), so we need to tell the link-checker to 
ignore that URL when looking for broken links.